requirements.txt
openai
pymupdf
typer[all]
python-dotenv
rich
